Fix builds on EPEL8 due to outdated components
Some of the new functionalities that are present in Fedora and are
needed for distgen to build using the refreshed build system are not
present in EPEL8.

This means that we have to rollback some changes to allow build on
EPEL8. This shouldn't impact functionality of the application, but
keeping it in the main repository would create mess in the codebase.
It can also be unpredictable how will setuptools work with different
config files, so it's better to just use one config.